re-implement multifile mod for performance
actually running multi file on redis was slooooooow, in particular when
flipping between files as the server/redis was source of truth.

Completely change implementation to use client side caching of files so
that flipping between different open files is more fluid.

Completely replace 'body' with openfiles manager on server and
bodyStore (a simple hash) on the client.

Much cleaner message passing between client and server